% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsl2sql.R
\name{createODK}
\alias{createODK}
\title{Create ODK form object}
\usage{
createODK(
  path,
  combine_views = data.table::data.table(),
  include_cols = NULL,
  change_names = NULL,
  exclude_cols = NULL,
  calc_types = NULL,
  calc_default = "text",
  type_default = "text"
)
}
\arguments{
\item{path}{Path to the ODK XLSForm (must be an xlsx file with the following sheets: survey, choices, settings)}

\item{combine_views}{\code{data.table} object used to configure questions that are in different repeats
but should be combined into the same column in a single view. Should be a data.table with the following columns:
\itemize{
\item \code{repeat_name_col}: the name of the column to be used to identify which repeat a value comes from.
\item \code{repeat_name}: Name of the repeat as programmed in the ODK XLSForm
\item \code{view}: Name of the view this question should be included in
\item \code{name_old}: Name of the question as programmed in the ODK XLSForm
\item \code{name_new}: Name of the column that should contain this question in the view
\item \code{repeat_id}: Name of the repeat as it should be entered in the column defined by \code{repeat_name_col}
}}

\item{include_cols}{Character vector mapping origin question names to views these questions should be included in.
Values should be \code{'always'} or the name of a repeat or view as defined in \code{combine_views} and/or the ODK XLSForm.}

\item{change_names}{Character vector mapping original question names to the names that should be used
for the SQL view column names.}

\item{exclude_cols}{Character vector listing the names of questions that should be excluded from the view.
Lines in the XLSForm with type note, or which begin/end a repeat or group are automatically excluded as separate variables from SQL views.}

\item{calc_types}{Character vector mapping the name of calculate fields to the type to be used in the SQL statement.
Calculate fields not present in this vector default to \code{calc_default}.}

\item{calc_default}{The default type to be used for calculate fields, when they do not appear in \code{calc_types}.}

\item{type_default}{The default type to be used for questions if data type cannot be guessed}
}
\value{
An S4 object of class \code{XLSForm}
}
\description{
This function creates an object containing all three tabs of the ODK XLSForm indicated,
calculates the data type for each variable, and splits the form into views based on
\code{combine_views} and/or the repeats in the ODK form. Returns an object of class \code{XLSForm}.
}
