% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsl2sql.R
\name{genCodebook,XLSForm-method}
\alias{genCodebook,XLSForm-method}
\alias{genCodebook}
\title{Generate codebook}
\usage{
\S4method{genCodebook}{XLSForm}(
  .Object,
  lang = "",
  includeCalculate = T,
  baseviewname = "",
  view_prefix = "",
  exclude = vector("character", 0),
  replace_choices = vector("character", 0),
  outfile = NULL
)
}
\arguments{
\item{.Object}{The XLSForm object}

\item{lang}{Language to use for creating the codebook.
Should be a two-letter language code used in the form to define labels/hints for different languages (e.g. EN for english)}

\item{baseviewname}{Name of the base view to use in the codebook (questions are organized by view).
Also used to identify questions by view, so must be equal to baseviewname as specified previously
(e.g. in \code{combine_views} argument to \code{\link[=create_odk]{create_odk()}})}

\item{view_prefix}{Prefix for the view name for labelling sections of the codebook}

\item{exclude}{Character vector of questions to exclude from the codebook}

\item{replace_choices}{Character vector mapping question names to strings which should be used in the codebook
instead of the options listed in the XLSForm. For example, useful when options are defined dynamically.}

\item{outfile}{File where the codebook should saved (must be a .txt file)}
}
\description{
Generate a codebook based on an ODK Form.
}
