% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsl2sql.R
\name{genPermissionsSQL,XLSForm-method}
\alias{genPermissionsSQL,XLSForm-method}
\alias{genPermissionsSQL}
\title{Generate query to change ownership of views}
\usage{
\S4method{genPermissionsSQL}{XLSForm}(.Object, schema, base_name, owner, view_prefix = "")
}
\arguments{
\item{.Object}{The XLSForm object to use for creating the queries}

\item{schema}{The schema in which views were created}

\item{base_name}{Name of the base view as created}

\item{owner}{What should the new owner of these views be (must be a user/group in the PostGreSQL database)}

\item{view_prefix}{The prefix added to view names as created}
}
\description{
This function generates a vector of SQL statements which change the ownership of the views defined by an XLSForm.
It should be called after generating SQL statements with \code{\link[=genSQL]{genSQL()}} and/or \code{\link[=genSQLAll]{genSQLAll()}}
and \emph{executing} these queries to create the views in the database.
}
