% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsl2sql.R
\name{genSQL,XLSForm-method}
\alias{genSQL,XLSForm-method}
\alias{genSQL}
\alias{genSQLAll,XLSForm-method}
\alias{genSQLAll}
\title{Generate SQL statement}
\usage{
\S4method{genSQL}{XLSForm}(
  .Object,
  schema,
  view_name,
  startdate,
  enddate,
  base = F,
  view_prefix = "",
  filter_version_sql = F,
  source_table = "openhds.form_submission",
  distinct_on = NULL,
  sm_exclude = vector("character", length = 0)
)

\S4method{genSQLAll}{XLSForm}(
  .Object,
  schema,
  view_name,
  startdate,
  enddate,
  base_view_name,
  view_prefix = "",
  filter_version_sql = F,
  source_table = "openhds.form_submission",
  distinct_on = NULL,
  sm_exclude = sm_exclude
)
}
\arguments{
\item{.Object}{\code{XLSForm} object to base the query on.}

\item{schema}{Name of the schema in which to create the view(s)}

\item{view_name}{Name of the view to create within \code{schema}.
For repeats this should match the name of the repeat or the \code{view} column in \code{combine_views} for \code{\link[=createODK]{createODK()}}.
Cannot be specified for \code{genSQLAll}, only for \code{genSQL}.}

\item{startdate}{The start date of the period for which forms should be included in this view}

\item{enddate}{The end date of the period for which forms should be included in this view}

\item{base}{Whether this is a call to generate the base query (\code{TRUE}) or a repeat query (\code{FALSE}).
Can only be specified in \code{genSQL}, but not in \code{genSQLAll}}

\item{view_prefix}{The prefix to add to the view name. Especially useful when creating views for repeats
to associate with the relevant form/activity}

\item{filter_version_sql}{Logical, indicating whether the view should use the form version supplied to filter.
If \code{TRUE}, \code{form_version} should be specified.}

\item{source_table}{The table which contains all form submissions from ODK.
This function assumes that one of the columns in this table (\code{as_json}) contains data in JSON format.}

\item{distinct_on}{Character vector indicating which question names (if any) should be used for
a \verb{DISTINCT ON} clause for the view. Useful when the same form has been sent to server multiple times by error.}

\item{sm_exclude}{Questions with type of \code{select_multiple} which should not be queried to create
separate columns for each possible response, and instead return the text string, possibly contining multiple responses.
Intended for use when the options for the question are defined dynamically so the choices sheet cannot contain all options.}

\item{base_view_name}{Name the view that should be created for the base statement
(ie. questions not in a repeat)}
}
\description{
Functions to generate SQL statements that create views for all repeats in the ODK form.
\code{genSQL} generates a single statement while \code{genSQLAll} is a convenience function
to generate the statements to create views for all parts of the form
(the base statement for questions outside repeats and a view for all repeats or as configured with \code{combine_views} in \code{\link[=createODK]{createODK()}}.
}
